pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swot-downup"
version = "0.1.0"
description = "Modular SWOT search/download/process/publish pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "SWOT Pipeline" }]
dependencies = [
  "pyyaml>=6.0",
  "typer>=0.12",
  "requests>=2.31",
  "numpy>=1.26",
  "xarray>=2024.1.0",
  "netCDF4>=1.6.5",
  "rasterio>=1.3.9",
  "shapely>=2.0.2",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "python-multipart>=0.0.9",
  "pyproj>=3.6.0",
  "pyshp>=2.3.1",
]

[project.optional-dependencies]
access = [
  "earthaccess>=0.11.0",
]
publish = [
  "google-cloud-storage>=2.14.0",
  "earthengine-api>=0.1.395",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "httpx>=0.27.0",
]

[project.scripts]
swot-pipeline = "swot_pipeline.cli:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["swot_pipeline*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
